NewTriggerInput created
